changes for real database

+++ database/dataHandler.js
@@ -3,7 +3,8 @@ const lodash = require('lodash')
 const demoProfiles = require(__dirname+'/demoData/profiles')
 const projects = require(__dirname+'/demoData/projects')
 const request = require('request');
-const fs = require('fs')
+const fs = require('fs');
+const { add } = require('lodash');
 
 function demoUser(token,callback){
   const User = demoProfiles.profiles.filter((user)=>user.personalData.userId == 2)
@@ -25,21 +26,37 @@ request.get(options, (err, resp, body) => {
     console.log("userId",JSON.parse(body).userId)
     const data = [User,JSON.parse(body)]
 
-    let rawData = fs.readFileSync(__dirname+'/data/users/users.json')
-    let users = JSON.parse(rawData)
-    users.users.push(JSON.parse(body))
-    fs.writeFileSync(__dirname+'/data/users/users.json',JSON.stringify(users))
-
-    rawData = fs.readFileSync(__dirname+'/data/users/users.json')
-    console.log(JSON.parse(rawData))
+    addUser(body)
+    
     callback(data)
 });
-
 }
-
 function demoProjects(callback){
   callback(projects.projects)
 }
+
+
+function addUser(body){
+  
+  let rawData = fs.readFileSync(__dirname+'/data/users/users.json')
+  let users = JSON.parse(rawData)
+  
+  let userExists = (users.users.filter((user)=>user.userId==JSON.parse(body).userId).length == 1)
+  if(userExists){
+    console.log("user already exists")
+    console.log(JSON.parse(body))
+  }else{
+    console.log("user added")
+    console.log(JSON.parse(body))
+    users.users.push(JSON.parse(body))
+    fs.writeFileSync(__dirname+'/data/users/users.json',JSON.stringify(users))
+  }
+}
+
+
+
+
+
 
 module.exports = {
   demoUser : demoUser,
